fix: move showDescription to own redux prop at root
