+++ src/reducers/selected.js
@@ -5,14 +5,12 @@ import { validateReducer } from '../modules/util'
 
 export const SET_SELECTED_ID = 'SET_SELECTED_ID'
 export const SET_SELECTED_ISLOADING = 'SET_SELECTED_ISLOADING'
-export const SET_SELECTED_SHOWDESCRIPTION = 'SET_SELECTED_SHOWDESCRIPTION'
 export const SET_SELECTED_ITEM_ACTIVE_TYPE = 'SET_SELECTED_ITEM_ACTIVE_TYPE'
 export const CLEAR_SELECTED_ITEM_ACTIVE_TYPES =
     'CLEAR_SELECTED_ITEM_ACTIVE_TYPES'
 
 export const DEFAULT_STATE_SELECTED_ID = null
 export const DEFAULT_STATE_SELECTED_ISLOADING = false
-export const DEFAULT_STATE_SELECTED_SHOWDESCRIPTION = false
 export const DEFAULT_STATE_SELECTED_ITEM_ACTIVE_TYPES = {}
 
 export const NON_EXISTING_DASHBOARD_ID = '0'
@@ -32,21 +30,6 @@ const isLoading = (state = DEFAULT_STATE_SELECTED_ISLOADING, action) => {
             return validateReducer(
                 action.value,
                 DEFAULT_STATE_SELECTED_ISLOADING
-            )
-        default:
-            return state
-    }
-}
-
-const showDescription = (
-    state = DEFAULT_STATE_SELECTED_SHOWDESCRIPTION,
-    action
-) => {
-    switch (action.type) {
-        case SET_SELECTED_SHOWDESCRIPTION:
-            return validateReducer(
-                action.value,
-                DEFAULT_STATE_SELECTED_SHOWDESCRIPTION
             )
         default:
             return state
@@ -75,7 +58,6 @@ const itemActiveTypes = (
 export default combineReducers({
     id,
     isLoading,
-    showDescription,
     itemActiveTypes,
 })
 
@@ -86,9 +68,6 @@ export const sGetSelectedRoot = state => state.selected
 export const sGetSelectedId = state => sGetSelectedRoot(state).id
 
 export const sGetSelectedIsLoading = state => sGetSelectedRoot(state).isLoading
-
-export const sGetSelectedShowDescription = state =>
-    sGetSelectedRoot(state).showDescription
 
 export const sGetSelectedItemActiveType = (state, id) =>
     sGetSelectedRoot(state).itemActiveTypes[id]
